requirements.txt
Flask==2.0.2
gunicorn==20.1.0
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==8.0.28
mysqlclient==2.1.0
numpy==1.20.3
os-win==5.4.0
pandas==1.2.4
python-dotenv==0.19.2
scikit-learn==1.0.2
flask-cors==3.0.10
urllib
uuid
boto3